requirements.txt
Flask
Flask-CORS
psycopg2-binary
PyJWT
Werkzeug